Add error and initial profile data parsing
Replicated the parsing error enum from the C++ code but also added some
contextual error info I think will be useful in future.

Implement parsing of the ProfileData and settle on a design for parsing
pointer specific widths. Need to just parse the array at the end but I'm
not yet sure how it gets the size for it