Bit of a tidy up
Given what I've learnt I've started rearranging some of things to
prepare to getting the output type. Added some constants I think I'll
need, and removed some previous thoughts sketched down in src/lib that are no
longer needed.

I've also created a binary for this with the aim of verifying by making
sure the llvm-profdata output and my output match. But with a nicer CLI
that makes it easy to discover the version of my tool