Start work on the hash table
This is a hash table stored on disk for the indexed instrumentation
profile. It's a very generic structure in LLVM but I think I only need
one concrete instantiation of it so I'll just be implementing the one I
need to start with. If necessary I can revisit and implement a generic
version