Add test case for broken merge and fix it (#16)
So the hash in the records is the function hash on raw profiles not the name hash so to correctly merge you need to apply the hash function to the name and if that hash is in merge it. But for text profiles the hash stored may be the name hash so you need to support both methods of merging the previous or the current. This PR does that to correct issues in instrumentation profile merges.